+++ src/analyzer.ts
@@ -7,7 +7,14 @@
 import * as vscode from 'vscode';
 import { spawnSync } from 'child_process';
 import { each } from 'lodash';
+const substituteVariables = require('var-expansion').substituteVariables; // no types available
+const slash = require('slash'); // no types available
 import * as pc from './paramcheck';
+
+let outputChannel: vscode.OutputChannel;
+export function setOutputChannel(channel: vscode.OutputChannel) {
+    outputChannel = channel;
+}
 
 export function runOnFile(config: {[key:string]:any}, fileName: string, workspaceDir: string) {
     if (!config['enable']) {
@@ -15,7 +22,7 @@ export function runOnFile(config: {[key:string]:any}, fileName: string, workspac
         return undefined;
     }
 
-    let params = getCppcheckParameters(config, false);
+    let params = getCppcheckParameters(config, false, true);
     params.push(fileName);
     let out = runCppcheck(params, config, workspaceDir);
     return out;
@@ -27,7 +34,20 @@ export function runOnWorkspace(config: {[key:string]:any}, workspaceDir: string)
         return undefined;
     }
 
-    let params = getCppcheckParameters(config, true);
+    let params = getCppcheckParameters(config, true, true);
+    params.push(workspaceDir);
+    let out = runCppcheck(params, config, workspaceDir);
+    return out;
+}
+
+/** The only difference between this and 'runOnWorkspace' is that the 'verbose' parameter is never used. */
+export function runLintMode(config: {[key:string]:any}, workspaceDir: string) {
+    if (!config['enable']) {
+        vscode.window.showInformationMessage('Cppcheck is not enabled.');
+        return undefined;
+    }
+
+    let params = getCppcheckParameters(config, true, false);
     params.push(workspaceDir);
     let out = runCppcheck(params, config, workspaceDir);
     return out;
@@ -39,11 +59,18 @@ function runCppcheck(params: string[], config: {[key:string]:any}, workspaceDir:
     let stdout = '' + result.stdout;
     let stderr = '' + result.stderr;
     let end = 'Cppcheck ended: ' + new Date().toString();
-    let out = [start, stdout, stderr, end].join('\n');
+    let resultsArray = [start, stdout, stderr, end];
+
+    if (config['outputCommandLine']) {
+        let commandLine = `${config['cppcheckPath']} ${params.join(' ')}`;
+        resultsArray.unshift(commandLine);
+    }
+
+    let out = resultsArray.join('\n');
     return out;
 }
 
-function getCppcheckParameters(config: {[key:string]:any}, unusedFunction: boolean) {
+function getCppcheckParameters(config: {[key:string]:any}, unusedFunction: boolean, useVerbose: boolean) {
     let enableParams = unusedFunction
                         ? [ '--enable=warning,style,performance,portability,information,unusedFunction' ]
                         : [ '--enable=warning,style,performance,portability,information' ];
@@ -61,7 +88,7 @@ function getCppcheckParameters(config: {[key:string]:any}, unusedFunction: boole
     let platformParams = getPlatformParams(config);
     params.push(platformParams);
 
-    if (config['verbose'] === true) {
+    if (useVerbose && config['verbose'] === true) {
         params.push('--verbose');
     }
 
@@ -74,11 +101,34 @@ function getIncludeParams(config: {[key:string]:any}) {
 
     if (paths) {
         each(paths, (element: string) => {
-            params.push('-I"' + element + '"');
+            let value = expandVariables(element);
+            if (value.error) {
+                outputChannel.appendLine(`Error expanding include path '${element}': ${value.error.message}`);
+            } else {
+                params.push(`-I"${value.result}"`);
+            }
         });
     }
 
     return params;
+}
+
+interface IExpansionResult {
+    error?: any;
+    result?: string;
+}
+
+function expandVariables(str: string): IExpansionResult {
+    process.env.workspaceRoot = vscode.workspace.rootPath;
+    let { value, error } = substituteVariables(str, { env: process.env });
+
+    if (error) {
+        return { error: error };
+    } else if (value === '') {
+        return { error: `Expanding '${str}' resulted in an empty string.` };
+    } else {
+        return { result: slash(value) };
+    }
 }
 
 function getPlatformParams(config: {[key:string]:any}) {
@@ -89,7 +139,7 @@ function getPlatformParams(config: {[key:string]:any}) {
             return '--platform=native';
         }
 
-        return '--platform=' + platform;
+        return `--platform=${platform}`;
     }
 
     return '--platform=native';
@@ -102,7 +152,7 @@ function getStandardParams(config: {[key:string]:any}) {
     if (standard) {
         each(standard, (element: string) => {
             if (pc.isValidStandard(element)) {
-                params.push('--std=' + element);
+                params.push(`--std=${element}`);
             }
         });
     }
@@ -120,7 +170,7 @@ function getDefineParams(config: {[key:string]:any}) {
 
     if (define) {
         each(define, (element: string) => {
-            params.push('-D' + element);
+            params.push(`-D${element}`);
         });
     }
 
@@ -133,7 +183,7 @@ function getUndefineParams(config: {[key:string]:any}) {
 
     if (undefine) {
         each(undefine, (element: string) => {
-            params.push('-U' + element);
+            params.push(`-U${element}`);
         });
     }
 
@@ -146,7 +196,7 @@ function getSuppressionParams(config: {[key:string]:any}) {
 
     if (suppressions) {
         each(suppressions, (element: string) => {
-            params.push('--suppress=' + element);
+            params.push(`--suppress=${element}`);
         });
     }